guardo cambios antes de crear la nueva rama
