+++ Backend/src/personas/models.py
@@ -6,11 +6,15 @@ from src.models import ModeloBase
 
 class Persona(ModeloBase):
     __tablename__ = "personas"
+
     id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
     nombre: Mapped[str] = mapped_column(String, index=True)
     apellido: Mapped[str] = mapped_column(String(30), index=False)
     legajo: Mapped[int] = mapped_column(Integer, index=False)
+    dni: Mapped[str] = mapped_column(String, unique=True, index=True)
     email: Mapped[str] = mapped_column(String, unique=True, index=True)
     rol_id: Mapped[int] = mapped_column(ForeignKey("roles.id"))
 
     rol: Mapped["Rol"] = relationship("Rol", back_populates="personas")
+    alumno: Mapped[Optional["Alumno"]] = relationship("Alumno", back_populates="persona", uselist=False)
+    docente: Mapped[Optional["Docente"]] = relationship("Docente", back_populates="persona", uselist=False)
