--- Frontend/src/styles/EncuestasCompletas.css
@@ -1,192 +1,0 @@
-/* Sección del encabezado */
-.header-section {
-    text-align: center;
-    margin-bottom: 2rem;
-}
-
-/* Cards de encuestas completadas horizontales */
-.encuesta-card-horizontal {
-    border: none;
-    border-radius: 12px;
-    box-shadow: 0 2px 8px #0000001a;
-    transition: all 0.3s ease;
-    background-color: #f8f9fa;
-}
-
-.encuesta-card-horizontal:hover {
-    transform: translateY(-2px);
-    box-shadow: 0 4px 15px #00000026;
-}
-
-.card-body-horizontal {
-    padding: 1.25rem;
-}
-
-/* Icono de encuesta completada */
-.encuesta-icon-completada {
-    width: 40px;
-    height: 40px;
-    background-color: #198754;
-    border-radius: 50%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin: 0 auto;
-}
-
-.encuesta-icon-completada i {
-    font-size: 1.2rem;
-    color: white;
-}
-
-/* Información de la encuesta */
-.encuesta-info {
-    padding-left: 0.5rem;
-}
-
-.asignatura-title-horizontal {
-    font-size: 1.1rem;
-    font-weight: 600;
-    color: #212529;
-    margin-bottom: 0.25rem;
-    line-height: 1.3;
-}
-
-.encuesta-meta {
-    font-size: 0.85rem;
-    color: #6c757d;
-}
-
-.profesor-text {
-    display: block;
-    margin-bottom: 0.25rem;
-}
-
-.fecha-completada {
-    display: flex;
-    align-items: center;
-    color: #495057;
-}
-
-.fecha-completada i {
-    color: #198754;
-    font-size: 0.9rem;
-}
-
-/* Badges */
-.badges-container {
-    text-align: center;
-    display: flex;
-    flex-wrap: wrap;
-    gap: 0.5rem;
-    justify-content: center;
-}
-
-.cursado-badge-horizontal {
-    font-size: 0.75rem;
-    font-weight: 600;
-    padding: 0.4rem 0.8rem;
-    border-radius: 8px;
-}
-
-.estado-badge-horizontal {
-    font-size: 0.7rem;
-    font-weight: 600;
-    padding: 0.3rem 0.6rem;
-    border-radius: 6px;
-}
-
-/* Botón de acción */
-.action-btn-horizontal {
-    font-size: 0.8rem;
-    font-weight: 600;
-    padding: 0.5rem 1rem;
-    border-radius: 8px;
-    width: 100%;
-    min-width: 120px;
-}
-
-/* Lista de encuestas */
-.encuestas-list {
-    max-width: 100%;
-}
-
-/* Estados de cargando y error */
-.loading-text {
-    font-size: 1.1rem;
-    color: #6c757d;
-    margin: 0;
-}
-
-.error-alert {
-    border-radius: 12px;
-    border: none;
-    box-shadow: 0 4px 12px #dc354533;
-}
-
-/* Estado vacío */
-.empty-state {
-    text-align: center;
-    padding: 3rem 1rem;
-    color: #6c757d;
-}
-
-.empty-icon {
-    font-size: 4rem;
-    margin-bottom: 1rem;
-    opacity: 0.5;
-}
-
-.empty-icon i {
-    color: #198754;
-}
-
-.empty-state h3 {
-    color: #495057;
-    margin-bottom: 1rem;
-}
-
-.empty-state p {
-    font-size: 1rem;
-    max-width: 400px;
-    margin: 0 auto;
-}
-
-/* Título principal */
-.titulo {
-    color: #212529;
-    font-weight: 700;
-    margin-bottom: 0.5rem;
-}
-
-.subtitulo {
-    color: #6c757d;
-    font-size: 1.1rem;
-    margin-bottom: 0;
-}
-
-/* Responsive para móviles */
-@media (max-width: 768px) {
-    .action-btn-horizontal {
-        width: 100%;
-        margin-top: 1rem;
-    }
-    
-    .badges-container {
-        text-align: left;
-        margin-top: 0.5rem;
-        justify-content: flex-start;
-    }
-    
-    .encuesta-icon-completada {
-        margin-bottom: 1rem;
-    }
-    
-    .card-body-horizontal {
-        padding: 1rem;
-    }
-    
-    .estado-badge-horizontal {
-        margin-top: 0.25rem;
-    }
-}
